First pass adding Room database Broke up MarvelResponse until seperate classes
