+++ app/src/main/java/com/laubetech/comiccovers/models/data/Comic.kt
@@ -1,0 +1,8 @@
+package com.laubetech.comiccovers.models.data
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class Comic(@PrimaryKey @ColumnInfo(name="issueId") val issueId:String,val issueTitle:String,val issueLength:String  )
