+++ app/src/main/java/com/laubetech/comiccovers/models/data/ComicDatabase.kt
@@ -1,0 +1,9 @@
+package com.laubetech.comiccovers.models.data
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(entities = arrayOf(Comic::class), version = 1)
+abstract class ComicDatabase : RoomDatabase(){
+    abstract fun comicDao(): ComicDao
+}
