+++ app/src/main/java/com/laubetech/comiccovers/models/data/ComicRepository.kt
@@ -1,0 +1,10 @@
+package com.laubetech.comiccovers.models.data
+
+class ComicRepository(private val comicDao: ComicDao){
+
+    suspend fun insert(comic: Comic){
+        comicDao.insertComic(comic)
+    }
+
+
+}
