+++ app/src/main/java/com/laubetech/comiccovers/models/response/Characters.kt
@@ -1,0 +1,8 @@
+package com.laubetech.comiccovers.models.response
+
+data class Characters (
+    val available: Long,
+    val collectionURI: String,
+    val items: List<Series>,
+    val returned: Long
+)
