+++ app/src/main/java/com/laubetech/comiccovers/models/response/Creators.kt
@@ -1,0 +1,8 @@
+package com.laubetech.comiccovers.models.response
+
+data class Creators (
+    val available: Long,
+    val collectionURI: String,
+    val items: List<CreatorsItem>,
+    val returned: Long
+)
