+++ app/src/main/java/com/laubetech/comiccovers/models/response/CreatorsItem.kt
@@ -1,0 +1,7 @@
+package com.laubetech.comiccovers.models.response
+
+data class CreatorsItem (
+    val name: String,
+    val resourceURI: String,
+    val role: String
+)
