+++ app/src/main/java/com/laubetech/comiccovers/models/response/Date.kt
@@ -1,0 +1,6 @@
+package com.laubetech.comiccovers.models.response
+
+data class Date (
+    val date: String,
+    val type: String
+)
