+++ app/src/main/java/com/laubetech/comiccovers/models/response/MarvelResponse.kt
@@ -1,0 +1,33 @@
+package com.laubetech.comiccovers.models.response
+
+data class MarvelResponse (
+    val attributionHTML: String,
+    val attributionText: String,
+    val code: Long,
+    val copyright: String,
+    val data: Data,
+    val etag: String,
+    val status: String
+)
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
