+++ app/src/main/java/com/laubetech/comiccovers/models/response/Price.kt
@@ -1,0 +1,6 @@
+package com.laubetech.comiccovers.models.response
+
+data class Price (
+    val price: Double,
+    val type: String
+)
