+++ app/src/main/java/com/laubetech/comiccovers/models/response/Series.kt
@@ -1,0 +1,6 @@
+package com.laubetech.comiccovers.models.response
+
+data class Series (
+    val name: String,
+    val resourceURI: String
+)
