+++ app/src/main/java/com/laubetech/comiccovers/models/response/Stories.kt
@@ -1,0 +1,8 @@
+package com.laubetech.comiccovers.models.response
+
+data class Stories (
+    val available: Long,
+    val collectionURI: String,
+    val items: List<StoriesItem>,
+    val returned: Long
+)
