+++ app/src/main/java/com/laubetech/comiccovers/models/response/StoriesItem.kt
@@ -1,0 +1,7 @@
+package com.laubetech.comiccovers.models.response
+
+data class StoriesItem (
+    val name: String,
+    val resourceURI: String,
+    val type: String
+)
