+++ app/src/main/java/com/laubetech/comiccovers/models/response/Thumbnail.kt
@@ -1,0 +1,6 @@
+package com.laubetech.comiccovers.models.response
+
+data class Thumbnail (
+    val extension: String,
+    val path: String
+)
