+++ app/src/main/java/com/laubetech/comiccovers/models/response/URL.kt
@@ -1,0 +1,6 @@
+package com.laubetech.comiccovers.models.response
+
+data class URL (
+    val type: String,
+    val url: String
+)
