+++ app/src/main/java/com/laubetech/comiccovers/ui/main/MainViewModel.kt
@@ -18,6 +18,8 @@ class MainViewModel : ViewModel() {
     init {
         reloadImage.value = false
         currentImageName.value = "image.jpg"
+
+
     }
 
     fun goButton( selectedComic:Long){
